test: move custom matchers to setup file
and add an additional toContainObject matcher
