+++ packages/server-wallet/jest/jest.config.js
@@ -4,7 +4,7 @@ const root = resolve(__dirname, '../');
 module.exports = {
   rootDir: root,
   collectCoverageFrom: ['src/**/*.{js,ts}'],
-  setupFilesAfterEnv: ['./jest/knex-setup-teardown.ts'],
+  setupFilesAfterEnv: ['./jest/custom-matchers.ts','./jest/knex-setup-teardown.ts'],
   testMatch: ['<rootDir>/**/__test__/**/?(*.)(spec|test).ts'],
   testEnvironment: 'node',
   testURL: 'http://localhost',
