+++ packages/server-wallet/src/__test__/create-and-close-channel/ledger-funding.test.ts
@@ -1,12 +1,5 @@
 import {CreateChannelParams} from '@statechannels/client-api-schema';
-import {
-  AllocationItem,
-  areAllocationItemsEqual,
-  BN,
-  makeAddress,
-  makeDestination,
-  Participant,
-} from '@statechannels/wallet-core';
+import {BN, makeAddress, makeDestination, Participant} from '@statechannels/wallet-core';
 import {ethers} from 'ethers';
 
 import {Outgoing} from '../..';
@@ -24,38 +17,6 @@ import {defaultTestConfig, overwriteConfigWithDatabaseConnection} from '../../co
 const ETH_ASSET_HOLDER_ADDRESS = makeAddress(ethers.constants.AddressZero);
 
 const tablesUsingLedgerChannels = ['channels', 'ledger_requests', 'ledger_proposals'];
-
-// TODO: Extending expect should be probably done in a setup file
-// This is probably fine for now since we only use toContainAllocationItem in these tests
-// This was moved from test-helpers.ts where it was a bit of footgun.
-// If you loaded the test-helpers.ts for any reason it would try to extend expect
-// which blows up if there is no expect defined
-expect.extend({
-  toContainAllocationItem(received: AllocationItem[], argument: AllocationItem) {
-    const pass = received.some(areAllocationItemsEqual.bind(null, argument));
-    if (pass) {
-      return {
-        pass: true,
-        message: () =>
-          `expected ${JSON.stringify(received, null, 2)} to not contain ${JSON.stringify(
-            argument,
-            null,
-            2
-          )}`,
-      };
-    } else {
-      return {
-        pass: false,
-        message: () =>
-          `expected ${JSON.stringify(received, null, 2)} to contain ${JSON.stringify(
-            argument,
-            null,
-            2
-          )}`,
-      };
-    }
-  },
-});
 
 let a = Wallet.create(
   overwriteConfigWithDatabaseConnection(defaultTestConfig(), {
