+++ packages/server-wallet/src/__test__/globals.d.ts
@@ -3,6 +3,7 @@ declare global {
   namespace jest {
     interface Matchers</* eslint-disable-line */ R> {
       toContainAllocationItem<R>(received: {amount: string; destination: string}): R;
+      toContainObject<R>(argument: unknown): R;
     }
   }
 }
